+++ voicevox_engine/preset/__init__.py
@@ -1,0 +1,1 @@
+from .PresetLoader import PresetLoader
